* Included pam_mkhomedir support in login role

* FIX nfsroot image and minor output format

* FIX minor style issues, included OpenSUSE Leap 42.2 template, improvements in rootfs generation

* FIX minor style issues, included OpenSUSE Leap 42.2 template, improvements in rootfs generation

* Included preliminary support for "set node" function
    Included cpu id map for Intel Broadwell
    Included Console options and Install repo support in the database

* Included a default INSTALL_REPO in the template config file

* Included root password update in the deploy time

* Enhancement in nodelist, new function to setup node description in the database, load PXE parameters from database per node, and minor style issues fixed

* unset the variable node after the loop

* FIX in help message

* FIX minor style issues, fix update_tools and update_configspace functions.

* FIX minor style issues

* Extended snow add node function

* snow remove also allows to remove a node list

* FIX minor issue in the template path in list_templates function

* Improved help info

* FIX minor style issues

* FIX minor style issues

* Merged node_list function as a part of boot function

* Included default image and template per node if defined in the database

* FIX minor style issue. The functions nreboot, npoweroff, nshutdown are now supporting a node list rather than just a range

* FIX issue in gnu parallel syntax

* included MAC and IP address in the node set

* included set_snow_json which abstrats the json file updates

* included set_snow_json which abstrats the json file updates

* included cpus, memory and disk size in the description of json file, as part of set_snow_json function

* included last_deploy in the description of json file, as part of set_snow_json function

* FIX default console options and install repo

* Minor change in a comment

* Initial role for torque

* FIX issue #63 - missing mail program and empty license variable

* FIX issue #63 - missing mail program and empty license variable

* minor tweek snow.conf-example

* FIX: issue #70 delay in the deploy when NFS is misconfigured

* FIX issue with legacy slurm runlevel daemon. Now it only uses slurmd.service

* Included new template with minimal packages

* Included new template with minimal packages

* FIX issue #11 #16 #55 #61

* FIX issues #58 #67 #72

* Included more comments in the snow.conf-example and renamed/removed some confusing variables.

* FIX issue #73 - updated help information in order to differentiate within poweroff and destroy

* FIX issue #67 - improved snow list error message

* Included error trapping in snow set node function

* Included logic for accommodating default node-based template

* minor fix in error message content

* Initial fusion of repos

* Functions clean-up and migrated/ported

* Install repo fusion ready for testing, created user_setup function

* Fixed issues with user definition and root password update when no password is provided

* Fix error while trying to copy some files not yet available

* Improved error message and excluded VIM custom configuration - hpcnow dotfiles

* Ready to be tested in the development branch

* README file updated

* FIX issue #3 - dom0 memory limits

* simplified git repo pulling from specifig branch

* update in default grub

* update in default grub

* included new function to replace text in files and updated virtualization install process to avoid duplicated lines in config fiels

* Fix package issues with ubuntu

* Fix minor issue in variable passing quotation in gawk function

* fix issue with get_distribution and force option in domain removal

* minor fix in eula download

* Several fixes in snow clone - NFSROOT type

* Improvements in nfsroot and squashfs support

* NFSROOT working with readonly rootfs and CentOS 7

* NFSROOT working with SuSE Leap. No support for overlayfs/unionfs in order to setup rm in memory

* Read-only NFSROOT image only working for CentOS/RHEL. Tuned dracut module is required to enable it for SuSE

* snow edited online with Bitbucket

* Included show nodes feature

* Included snow list roles function. The style of the role scripts are normalised

* By default it shows all the compute nodes

* By default it shows all the compute nodes

* fixed issues with leap deployment and included chroot enviroment function

* removed a comment

* Fix authentication on domains by installing libpam-sss and libnss-sss

* Fix user home creation on login

* Fix permissions and owner for sssd.conf to be able to start the daemon

* Updated sNow version in the install.sh

* Fix interactive from login node

* Fix eb_wrap path to slurm.env

* bin folder clean-up

* Fix user home creation on login

* Add skel and ACL on user home creation

* Minor fixes in list format and paths of images

* Fix issue in cfs role: missing prefix in the header

* Fix false positive in the mountpoint warning during the installation

* Fix: NET_LLF is optional

* Fix minor issues with deploy template and error message in snow init

* Fix enable public ssh instance in login role

* Fix included default values for sNow and HPCNow users, by default HPCNow_Support=none, so HPCNow user is not created by default

* improved Easybuild and Lmod reliability for interrupted deployments

* roll back to hooks path evaluation

* Fix issues related with services under ro nfsroot

* Included new functions to create raids and filesystems. Included an example of first_boot_hook

* Added two examples for building local file system as a hook

* Improved eb_wrap and interactive. Fairsharing is not longer default requirement. AccountingStorageEnforce=nojobs

* Report error when snow.conf is not available and snow CLI is used. Added spinner in the snow shutdown domains and error tracking

* Included check_host_status in elegible power management functions

* Fixed issue with PDSH protocol

* Integrated first boot hooks in the image system, fix issues with mutable files in ro nfsroot and disabled readahead in CentOS

* Readahead operations disabled in the centos template

* Documentation improved in snow.conf-example: HPCNow_Support=none by default and hpcnow users not created by default. warning message after snow init when the snow server is also NFS server.

* extended help message for shutdown

* Fix wrong message in /etc/issue and /etc/issue.net

* Improved snow.conf-example documentation (LLF is optional)

* Patch network config in centos with linkdelay=20 in order to fix issues with some 10Gb devices

* Destroy function also allows to manage all the domains and clusters of compute nodes

* Removed System Imager from deploy role, as not longer required

* Fix issue #100 (updates in the DB after a deploy). Fix IPv6 issues related with RPCbind Server Activation Socket. Included memtest as image

* Improved error message in destroy_cluster function and removed boot_delay in the boot process (only required in deploy process)

* Minor fix in network availability in deployment stage

* fix issue with sssd dependencies and fixed minor code style issue in patch_network_configuration function

* Fix issue related with first boot hooks

+++ share/common.d/users_setup.sh
@@ -3,46 +3,72 @@
 # Developed by Jordi Blasco <jordi.blasco@hpcnow.com>
 # For more information, visit the official website : www.hpcnow.com/snow
 #
-function setup_snow_user()
+
+
+function setup_user()
 {
-    # Check UIDs and GIDs
-    if [[ -z $(getent passwd $sNow_USER) ]]; then
-        groupadd -g $sNow_GID $sNow_GROUP
-        useradd -u $sNow_UID -g $sNow_GID -c "sNow! Admin User" -s /bin/bash -d $SNOW_HOME/$sNow_USER  $sNow_USER 
-    elif [[ "$(id -u $sNow_USER)" != "$sNow_UID"  &&  "$(id -g $sNow_USER)" != "$sNow_GID" ]]; then
-        groupmod -g $sNow_GID $sNow_GROUP
-        usermod -u $sNow_UID -g $sNow_GID $sNow_USER
-        usermod -c "sNow! Admin User" -g $sNow_GID -d $SNOW_HOME/$sNow_USER -s /bin/bash -m -u $sNow_UID $sNow_USER
+    local user_name=$1
+    local user_uid=$2
+    local user_gid=$3
+    local user_group=$4
+    local user_sudoers=$5
+    local user_shell=$6
+    local user_gecos="$7"
+
+    if [[ $# < 7 ]]; then
+        error_exit "User ${user_name} can not be created because some parameters are missing."
     fi
-    # Don't require password for sNow! Admin user sudo
-    echo "$sNow_USER ALL=(ALL) NOPASSWD: ALL" >> /etc/sudoers
-    if is_golden_node; then
-        # Configure public key auth for the sNow! Admin User
-        mkdir -p $SNOW_HOME/$sNow_USER/.ssh
-        # Setup the ACLs to the right user
-        chown -R $sNow_USER:$sNow_GROUP $SNOW_HOME/$sNow_USER
+
+    if [[ -z $(getent passwd ${user_name}) ]]; then
+        bkp /etc/group
+        bkp /etc/passwd
+        bkp /etc/shadow
+        groupadd -g ${user_gid} ${user_group}
+        useradd -u ${user_uid} -g ${user_gid} -c "${user_gecos}" -s ${user_shell} -d $SNOW_HOME/${user_name}  ${user_name} 
+    elif [[ "$(id -u ${user_name})" != "${user_uid}"  &&  "$(id -g ${user_name})" != "${user_gid}" ]]; then
+        groupmod -g ${user_gid} ${user_group}
+        usermod -u ${user_uid} -g ${user_gid} ${user_name}
+        usermod -c "${user_gecos}" -g ${user_gid} -d $SNOW_HOME/${user_name} -s ${user_shell} -m -u ${user_uid} ${user_name}
+    fi
+
+    if [[ "${user_sudoers}" == "true" ]]; then
+        bkp /etc/sudoers
+        echo "${user_name} ALL=(ALL) ALL" >> /etc/sudoers
+    fi
+
+    if [[ "${user_sudoers}" == "nopasswd" ]]; then
+        bkp /etc/sudoers
+        echo "${user_name} ALL=(ALL) NOPASSWD: ALL" >> /etc/sudoers
+    fi
+
+    if [[ ! -f $SNOW_HOME/${user_name}/.ssh/id_rsa ]]; then
+        if [[ ! -e $SNOW_HOME/${user_name}/.ssh ]]; then
+            mkdir -p $SNOW_HOME/${user_name}/.ssh
+            chown -R ${user_name}:${user_group} $SNOW_HOME/${user_name}/.ssh
+        fi
+        sudo -u ${user_name} ssh-keygen -t rsa -f $SNOW_HOME/${user_name}/.ssh/id_rsa -q -P ""
+        bkp $SNOW_HOME/${user_name}/.ssh/authorized_keys
+        cat $SNOW_HOME/${user_name}/.ssh/id_rsa.pub > $SNOW_HOME/${user_name}/.ssh/authorized_keys
+        echo "Host *" > $SNOW_HOME/${user_name}/.ssh/config
+        echo "    StrictHostKeyChecking no" >> $SNOW_HOME/${user_name}/.ssh/config
+        echo "    UserKnownHostsFile /dev/null" >> $SNOW_HOME/${user_name}/.ssh/config
+        echo "    PasswordAuthentication no" >> $SNOW_HOME/${user_name}/.ssh/config
+        chown ${user_name}:${user_group} $SNOW_HOME/${user_name}/.ssh/authorized_keys
+        chown ${user_name}:${user_group} $SNOW_HOME/${user_name}/.ssh/config
     fi
 } 1>>$LOGFILE 2>&1
-
-function setup_hpcnow_user()
-{
-    # Check UIDs and GIDs
-    if [[ -z $(getent passwd $HPCNow_USER) ]]; then
-        useradd -u $HPCNow_UID -g $HPCNow_GID -c "HPCNow! User" -s /bin/bash -d $SNOW_HOME/$HPCNow_USER  $HPCNow_USER 
-    elif [[ "$(id -u $HPCNow_USER)" != "$HPCNow_UID"  &&  "$(id -g $HPCNow_USER)" != "$HPCNow_GID" ]]; then
-        usermod -c "HPCNow! User" -g $HPCNow_GID -d $SNOW_HOME/$HPCNow_USER -s /bin/bash -m -u $HPCNow_UID $HPCNow_USER
-    fi
-    # Don't require password for sNow! Admin user sudo
-    echo "$HPCNow_USER ALL=(ALL) NOPASSWD: ALL" >> /etc/sudoers
-} 1>>$LOGFILE 2>&1
-
 
 function setup_ssh()
 {
-    setup_snow_user
-    if [[ "$HPCNow_Support" != "none" ]]; then
-        setup_hpcnow_user
+    # Users Setup
+    if [[ ! -z "${MASTER_PASSWORD}" ]]; then
+        echo "root:${MASTER_PASSWORD}" | chpasswd
     fi
+    setup_user "${sNow_USER}" "${sNow_UID}" "${sNow_GID}" "${sNow_GROUP}" "nopasswd" "/bin/bash" "sNow! Admin User"
+    if [[ "${HPCNow_Support}" != "none" ]]; then
+        setup_user "${HPCNow_USER}" "${HPCNow_UID}" "${HPCNow_GID}" "${HPCNow_GROUP}" "nopasswd" "/bin/bash" "HPCNow! Admin User"
+    fi
+    # Setup SSH
     mkdir -p /root/.ssh
     cp -p $SNOW_HOME/$sNow_USER/.ssh/authorized_keys /root/.ssh/authorized_keys
     # Allow support for GPFS requirements
@@ -51,16 +77,23 @@ function setup_ssh()
     chown -R root:root /root/.ssh
     chmod 700 /root/.ssh
     chmod 640 /root/.ssh/authorized_keys
+    chmod 400 /root/.ssh/id_rsa
     cp -pr $SNOW_CONF/system_files/etc/ssh/ssh_host_* /etc/ssh/
-    cp -p $SNOW_CONF/system_files/etc/ssh/shosts.equiv /etc/ssh/
-    cp -p $SNOW_CONF/system_files/etc/ssh/ssh_known_hosts /etc/ssh/ssh_known_hosts
+    if [[ -e $SNOW_CONF/system_files/etc/ssh/shosts.equiv ]]; then
+        cp -p $SNOW_CONF/system_files/etc/ssh/shosts.equiv /etc/ssh/
+    fi
+    if [[ -e $SNOW_CONF/system_files/etc/ssh/ssh_known_hosts ]]; then
+        cp -p $SNOW_CONF/system_files/etc/ssh/ssh_known_hosts /etc/ssh/ssh_known_hosts
+    fi
     if [[ -e /usr/lib64/ssh/ssh-keysign ]]; then
         chmod u+s /usr/lib64/ssh/ssh-keysign
     fi
     if [[ -e /usr/libexec/openssh/ssh-keysign ]]; then
         chmod u+s /usr/libexec/openssh/ssh-keysign
     fi
-    cp -p /etc/ssh/shosts.equiv /root/.shosts
+    if [[ -e /etc/ssh/shosts.equiv ]]; then
+        cp -p /etc/ssh/shosts.equiv /root/.shosts
+    fi
     sed -i "s/RhostsRSAAuthentication no/RhostsRSAAuthentication yes/g" /etc/ssh/sshd_config
     sed -i "s/HostbasedAuthentication no/HostbasedAuthentication yes/g" /etc/ssh/sshd_config
     sed -i "s/#   StrictHostKeyChecking ask/StrictHostKeyChecking no/g" /etc/ssh/ssh_config
